Add some integer casts to prevent overflows on 64bit Windows.
These were caught by MSVC warnings (when compiling in 64bit mode).

+++ tests/data_structures/semistatic_graph.cpp
@@ -26,19 +26,19 @@ using Graph = SemistaticGraph<int, const char*>;
 using node_iterator = Graph::node_iterator;
 using edge_iterator = Graph::edge_iterator;
 
-vector<size_t> no_neighbors{};
+vector<int> no_neighbors{};
 
 struct SimpleNode {
-  size_t id;
+  int id;
   const char* value;
-  const vector<size_t>* neighbors;
+  const vector<int>* neighbors;
   bool is_terminal;
   
-  size_t getId() { return id; }
+  int getId() { return id; }
   const char* getValue() { return value; }
   bool isTerminal() { return is_terminal; }
-  vector<size_t>::const_iterator getEdgesBegin() { return neighbors->begin(); }
-  vector<size_t>::const_iterator getEdgesEnd() { return neighbors->end(); }
+  vector<int>::const_iterator getEdgesBegin() { return neighbors->begin(); }
+  vector<int>::const_iterator getEdgesEnd() { return neighbors->end(); }
 };
 
 
@@ -88,7 +88,7 @@ void test_1_node_no_edges_terminal() {
 }
 
 void test_1_node_self_edge() {
-  vector<size_t> neighbors = {2};
+  vector<int> neighbors = {2};
   vector<SimpleNode> values{{2, "foo", &neighbors, false}};
   Graph graph(values.begin(), values.end(), -1, -2);
   Assert(graph.find(0) == graph.end());
@@ -108,7 +108,7 @@ void test_1_node_self_edge() {
 }
 
 void test_2_nodes_one_edge() {
-  vector<size_t> neighbors = {2};
+  vector<int> neighbors = {2};
   vector<SimpleNode> values{{2, "foo", &no_neighbors, false}, {3, "bar", &neighbors, false}};
   Graph graph(values.begin(), values.end(), -1, -2);
   Assert(graph.find(0) == graph.end());
@@ -132,7 +132,7 @@ void test_2_nodes_one_edge() {
 }
 
 void test_3_nodes_two_edges() {
-  vector<size_t> neighbors = {2, 4};
+  vector<int> neighbors = {2, 4};
   vector<SimpleNode> values{{2, "foo", &no_neighbors, false}, {3, "bar", &neighbors, false}, {4, "baz", &no_neighbors, true}};
   Graph graph(values.begin(), values.end(), -1, -2);
   Assert(graph.find(0) == graph.end());
@@ -165,7 +165,7 @@ void test_3_nodes_two_edges() {
 void test_add_node() {
   vector<SimpleNode> old_values{{2, "foo", &no_neighbors, false}, {4, "baz", &no_neighbors, true}};
   Graph old_graph(old_values.begin(), old_values.end(), -1, -2);
-  vector<size_t> neighbors = {2, 4};
+  vector<int> neighbors = {2, 4};
   vector<SimpleNode> new_values{{3, "bar", &neighbors, false}};
   Graph graph(old_graph, new_values.begin(), new_values.end());
   Assert(graph.find(0) == graph.end());
@@ -196,7 +196,7 @@ void test_add_node() {
 }
 
 void test_set_terminal() {
-  vector<size_t> neighbors = {2, 4};
+  vector<int> neighbors = {2, 4};
   vector<SimpleNode> values{{2, "foo", &no_neighbors, false}, {3, "bar", &neighbors, false}, {4, "baz", &no_neighbors, true}};
   Graph graph(values.begin(), values.end(), -1, -2);
   graph.find(3).setTerminal();
@@ -219,7 +219,7 @@ void test_set_terminal() {
 }
 
 void test_move_constructor() {
-  vector<size_t> neighbors = {2};
+  vector<int> neighbors = {2};
   vector<SimpleNode> values{{2, "foo", &no_neighbors, false}, {3, "bar", &neighbors, false}};
   Graph graph1(values.begin(), values.end(), -1, -2);
   Graph graph = std::move(graph1);
@@ -237,7 +237,7 @@ void test_move_constructor() {
 }
 
 void test_move_assignment() {
-  vector<size_t> neighbors = {2};
+  vector<int> neighbors = {2};
   vector<SimpleNode> values{{2, "foo", &no_neighbors, false}, {3, "bar", &neighbors, false}};
   Graph graph1(values.begin(), values.end(), -1, -2);
   Graph graph;
@@ -256,7 +256,7 @@ void test_move_assignment() {
 }
 
 void test_incomplete_graph() {
-  vector<size_t> neighbors = {2};
+  vector<int> neighbors = {2};
   vector<SimpleNode> values{{1, "foo", &neighbors, false}};
 
   Graph graph(values.begin(), values.end(), -1, -2);
